Switched to current original casadi.
The current casadi version does no longer provide a octave interface
using swig. That's why we use the friedrichatgc fork (on github) till now.
This branch switches back to the current casadi version and added a
octave swig interface for casadi to MBXMLUtils. Only the, for
mbxmlutilspp requried, SX and DMatrix casadi classes are supported.