Fixed issue #29
"Move camera with this body" does not work once a body was disabled and
enabled again.

Please note that if the camera is moving with a body then any change on
this body will release the camera from this body (by design) but now the
camera can be reenabled to move with this body.