Added cast to casadi::SX in Eval.
This cast also works for interpreter values being not a SWIG type but a
normal matrix in the interpreter language.
This is required to allow constant expressions in the interpreter to be
interpreted as a casadi::SX symbolic expression.

This was already working but was lost during the Eval class rework.
A test for this is added in the next MBSim commit (newUpdateConcept branch)