Replaced the value type of the Eval class
This type was std::shared_ptr<void> and stored everything as a "void*"
which represents the swig type of the script language.
Now this type is boost::variant<std::shared_ptr<void>, casadi::SXFunction>
and stores a casadi::SXFunction directly and all other types as a
"void*" which represents the swig type of the script language.

This is possible since casadi::SXFunction is not required inside the
script language.
This allows to remove the type casadi::SXFunction from the swig
generation.
This simplifies e.g. the octave swig generation a lot since the
complicated casadi::SXFunction must not be generated anymore.