Added a new simpletype "PV%script".
If an element has this type then the text content of this element is not
modified by the preprocessor but a processing instruction element
"ScriptParameter" is added. This PI contains all current parameters (one
per line) in the xmlflateval notation (type:parname=value).
This may be used to parametrize arbitary scripts which are unknown by
the preprocessor but known by the real program (mbsimxml).

It is used in MBSim to initialize mbsim module object written in python.