Enabled external XPath based parameters for all Evaluators
Before this change such a par was converted using CodeString (to flatxml
syntax) and stored in the XML tree and later evaluatored using the
evaluator which fails if the evaluators does not understand flatxml
syntax (e.g. python). Now such values are stored as user data on the
DOMElement as a Eval::Value*. This enabled all evaluators and avoid the
conversion to string and reconvertion to Eval::Value.