Fixed issue #32
Include the bbox of any object in the bbox calculation. This ensures the
the bbox of any body, which is 1.05 larger than its bounds, is not
clipped by the cameras clipping planes.
This has the negative effect that enabling the bbox of any object
enlarges that object by 1.05. Hence, the bbox of any parent object are
also enlarged (this hold recursively up to the root).
However, this harms less then the other issue.