Fixed threaded edge calculation
It seem to be illegal to add to a So* object from a different thread
even if the So* object is not part of the scene graph till now.
Hence, store everything in std containers during the threaded operation
and copy it afterward to a So* object.

This, hopefully, fixes sporadic crashes of openmbv and mbsimgui
(especially under high load during the nightly builds)