Export all symbols from classes needed by MBSimGUI.
Due to some unknown reason it was already required to explicitly export
symbols from openmbv which are required by mbsimgui on Windows.
Now also the Qt staticMetaObject is not exported due to some unknown
reason.
It not possible to export staticMetaObject explicity since it is defined
by Q_OBJECT. Hence, Qt exports always all symbols from a class and uses
the PImpl idiom to encapsulate private parts.
Hence, switch to export of all symbols of the classes (but still to not
use PImpl).