Workaround for a octave bug!?
Deleting local variables from the octave interpreter seem not to work
reliable. Every second time all (not only local) variables seems to be
removed. This seems to be a bug in octave.
This is now worked around by build the list of local variables as the
difference of all variables minus the global variables.
All these variables are removed than.

This should fix the errors im MBSimGUI.