Improved performance
When many objects are loaded the performance was quite bad due to the
checking of clone objects.
The check if a new object is a clone object is now optimized to improve
load time performance (also also unload time performacne).

(cloneing by be removed anyway in the future since this implementation
 is quite strange and a OpenMB"Viewer" must not support the edit feature
 which is the only reason why the cloneing exists)