Reworked the handling of embed processing information
All embed processing information are stored now as userdata and
written to XML PIs BEFORE the element if such a DOM is serialized.
Writing it before is done to be XML conform even for empty content
elements which are even not allowed to have child PIs or child comments
or empty (formatting) text nodes.

Moreover:
- the preprocessor does not a validation after processAndGetDocument to
  ensure that the final preprocessed document is valid (all embeda are correct)
  Validation by the caller of processAndGetDocument is removed.
- Document normalize is not longer needed and removed.
- serialize is now done always with prettyPrint