Convert the Windows WM_CLOSE signal to a posix SIGTERM signal
On Windows one cannot send SIGTERM or SIGINT to a process, only the
process itself can send such a signal to itself.
Interprocess signal handling is done on Windows using window manager
signals. Such signals are handled propably e.g. by Qt and convert to a
corresponding Qt event. But for a console application we need to handle
this by our own.
windows_signal_conversion.cc was introduced for this. It creates a
hidden dummy window in a new thread and listens for window manger events
in this thread. If a WM_CLOSE event is detected (usually send from
another process) a SIGTERM signal is raised (this is possible see
above). This way a WM_CLOSE event is handled analog to a posix SIGTERM
signal and the platform independent action for a SIGTERM happens.

This all is required to react on Windows to e.g. user exit requests
which are send to the process via the WM_CLOSE event e.g. by
QProcess::terminate().